Fix a bug in 31445a56 and 192cf92
This bug prevented players from starting faction vehicles as some functions still used the faction's type to check for unauthorized entry. This bugfix makes sure that the script checks for the faction ID instead of the type.

Also, this bug fixes an issue with job vehicles that made it impossible to start them.